temporary for #10
Given that the site is live, I recommend removing the links until they
are go somewhere. The site is cleaner and less confusing without the
dead end links.

Per issue #13, it would be better (IMHO) to be using Liquid comments
that do not appear in the rendered pages. I am using HTML comments to
be consistent with comments elsewhere.

Signed-off-by: Grant Pax <grant@fwep.twostewards.com>