swapping content of 400-run with 700-deploy
this is suggested but not necessary update. Changes is suggested because
this update follows the Quick Start Tutorial linear approach and makes
the flow of API-Provider-Application without disruption. Also removes
additional gogo command test:eval that was introduced in the earlier
version. I believe eval:eval and test:eval was introducing noise to
otherwise perfect linear flow. Please also consider the commit to
700--deploy page before reviewing this commit.

Signed-off-by: Daghan Acay <daghanacay@gmail.com>