Typo: UserApplication should be UpperApplication
I'm assuming that this was a typo - as there's no UserApplication and later the text continues to talk about the UpperApplication. Tiny glitch - I somewhat "autocorrected" it on first read, but figured that something was slightly irritating. Took me a bit to figure it out - that's how tiny it is